fix(Private): rename private folders to internal BREAKING CHANGE: change namespace detail in private folders to internal namespace
