+++ include/geode/io/mesh/internal/dxf_input.h
@@ -23,37 +23,34 @@
 
 #pragma once
 
-#include <string>
-#include <vector>
-
-#include <geode/mesh/io/triangulated_surface_output.h>
+#include <geode/mesh/io/polygonal_surface_input.h>
 
 namespace geode
 {
-    FORWARD_DECLARATION_DIMENSION_CLASS( TriangulatedSurface );
-    ALIAS_2D( TriangulatedSurface );
+    FORWARD_DECLARATION_DIMENSION_CLASS( PolygonalSurface );
+    ALIAS_3D( PolygonalSurface );
 } // namespace geode
 
 namespace geode
 {
-    namespace detail
+    namespace internal
     {
-        class TriangleOutput final : public TriangulatedSurfaceOutput< 2 >
+        class DXFInput final : public PolygonalSurfaceInput< 3 >
         {
         public:
-            explicit TriangleOutput( std::string_view filename )
-                : TriangulatedSurfaceOutput< 2 >( filename )
+            explicit DXFInput( std::string_view filename )
+                : PolygonalSurfaceInput< 3 >( filename )
             {
             }
 
             static std::string_view extension()
             {
-                static constexpr auto ext = "triangle";
+                static constexpr auto ext = "dxf";
                 return ext;
             }
 
-            std::vector< std::string > write(
-                const TriangulatedSurface2D &surface ) const final;
+            std::unique_ptr< PolygonalSurface3D > read(
+                const MeshImpl& impl ) final;
         };
-    } // namespace detail
+    } // namespace internal
 } // namespace geode
