+++ include/geode/io/mesh/internal/obj_triangulated_output.h
@@ -23,34 +23,38 @@
 
 #pragma once
 
-#include <geode/mesh/io/polygonal_surface_input.h>
+#include <string>
+#include <vector>
+
+#include <geode/mesh/io/triangulated_surface_output.h>
 
 namespace geode
 {
-    FORWARD_DECLARATION_DIMENSION_CLASS( PolygonalSurface );
-    ALIAS_3D( PolygonalSurface );
+    FORWARD_DECLARATION_DIMENSION_CLASS( TriangulatedSurface );
+    ALIAS_3D( TriangulatedSurface );
 } // namespace geode
 
 namespace geode
 {
-    namespace detail
+    namespace internal
     {
-        class PLYInput final : public PolygonalSurfaceInput< 3 >
+        class OBJTriangulatedOutput final
+            : public TriangulatedSurfaceOutput< 3 >
         {
         public:
-            explicit PLYInput( std::string_view filename )
-                : PolygonalSurfaceInput< 3 >( filename )
+            explicit OBJTriangulatedOutput( std::string_view filename )
+                : TriangulatedSurfaceOutput< 3 >( filename )
             {
             }
 
             static std::string_view extension()
             {
-                static constexpr auto ext = "ply";
+                static constexpr auto ext = "obj";
                 return ext;
             }
 
-            std::unique_ptr< PolygonalSurface3D > read(
-                const MeshImpl& impl ) final;
+            std::vector< std::string > write(
+                const TriangulatedSurface3D &surface ) const final;
         };
-    } // namespace detail
+    } // namespace internal
 } // namespace geode
