+++ include/geode/io/mesh/internal/ply_input.h
@@ -33,24 +33,24 @@ namespace geode
 
 namespace geode
 {
-    namespace detail
+    namespace internal
     {
-        class DXFInput final : public PolygonalSurfaceInput< 3 >
+        class PLYInput final : public PolygonalSurfaceInput< 3 >
         {
         public:
-            explicit DXFInput( std::string_view filename )
+            explicit PLYInput( std::string_view filename )
                 : PolygonalSurfaceInput< 3 >( filename )
             {
             }
 
             static std::string_view extension()
             {
-                static constexpr auto ext = "dxf";
+                static constexpr auto ext = "ply";
                 return ext;
             }
 
             std::unique_ptr< PolygonalSurface3D > read(
                 const MeshImpl& impl ) final;
         };
-    } // namespace detail
+    } // namespace internal
 } // namespace geode
