+++ include/geode/io/mesh/internal/stl_input.h
@@ -23,37 +23,34 @@
 
 #pragma once
 
-#include <string>
-#include <vector>
-
-#include <geode/mesh/io/polygonal_surface_output.h>
+#include <geode/mesh/io/triangulated_surface_input.h>
 
 namespace geode
 {
-    FORWARD_DECLARATION_DIMENSION_CLASS( PolygonalSurface );
-    ALIAS_3D( PolygonalSurface );
+    FORWARD_DECLARATION_DIMENSION_CLASS( TriangulatedSurface );
+    ALIAS_3D( TriangulatedSurface );
 } // namespace geode
 
 namespace geode
 {
-    namespace detail
+    namespace internal
     {
-        class OBJPolygonalOutput final : public PolygonalSurfaceOutput< 3 >
+        class STLInput final : public TriangulatedSurfaceInput< 3 >
         {
         public:
-            explicit OBJPolygonalOutput( std::string_view filename )
-                : PolygonalSurfaceOutput< 3 >( filename )
+            explicit STLInput( std::string_view filename )
+                : TriangulatedSurfaceInput< 3 >( filename )
             {
             }
 
             static std::string_view extension()
             {
-                static constexpr auto ext = "obj";
+                static constexpr auto ext = "stl";
                 return ext;
             }
 
-            std::vector< std::string > write(
-                const PolygonalSurface3D &surface ) const final;
+            std::unique_ptr< TriangulatedSurface3D > read(
+                const MeshImpl& impl ) final;
         };
-    } // namespace detail
+    } // namespace internal
 } // namespace geode
