+++ include/geode/io/model/internal/msh_output.h
@@ -23,34 +23,32 @@
 
 #pragma once
 
-#include <geode/mesh/io/triangulated_surface_input.h>
+#include <string>
+#include <vector>
+
+#include <geode/model/representation/io/brep_output.h>
 
 namespace geode
 {
-    FORWARD_DECLARATION_DIMENSION_CLASS( TriangulatedSurface );
-    ALIAS_3D( TriangulatedSurface );
-} // namespace geode
-
-namespace geode
-{
-    namespace detail
+    namespace internal
     {
-        class STLInput final : public TriangulatedSurfaceInput< 3 >
+        class MSHOutput final : public BRepOutput
         {
         public:
-            explicit STLInput( std::string_view filename )
-                : TriangulatedSurfaceInput< 3 >( filename )
+            explicit MSHOutput( std::string_view filename )
+                : BRepOutput( filename )
             {
             }
 
             static std::string_view extension()
             {
-                static constexpr auto ext = "stl";
+                static constexpr auto ext = "msh";
                 return ext;
             }
 
-            std::unique_ptr< TriangulatedSurface3D > read(
-                const MeshImpl& impl ) final;
+            std::vector< std::string > write( const BRep& brep ) const final;
+
+            bool is_saveable( const BRep& brep ) const final;
         };
-    } // namespace detail
+    } // namespace internal
 } // namespace geode
