+++ src/geode/io/mesh/CMakeLists.txt
@@ -77,20 +77,20 @@ add_geode_library(
         "detail/vtu_tetrahedral_input.h"
         "detail/vtu_tetrahedral_output.h"
     PRIVATE_HEADERS
-        "private/assimp_input.h"
-        "private/assimp_output.h"
-        "private/dxf_input.h"
-        "private/obj_input.h"
-        "private/obj_polygonal_output.h"
-        "private/obj_triangulated_output.h"
-        "private/ply_input.h"
-        "private/ply_output.h"
-        "private/smesh_curve_input.h"
-        "private/smesh_input.h"
-        "private/smesh_triangulated_input.h"
-        "private/stl_input.h"
-        "private/stl_output.h"
-        "private/triangle_output.h"
+        "internal/assimp_input.h"
+        "internal/assimp_output.h"
+        "internal/dxf_input.h"
+        "internal/obj_input.h"
+        "internal/obj_polygonal_output.h"
+        "internal/obj_triangulated_output.h"
+        "internal/ply_input.h"
+        "internal/ply_output.h"
+        "internal/smesh_curve_input.h"
+        "internal/smesh_input.h"
+        "internal/smesh_triangulated_input.h"
+        "internal/stl_input.h"
+        "internal/stl_output.h"
+        "internal/triangle_output.h"
     PUBLIC_DEPENDENCIES
         OpenGeode::image
     PRIVATE_DEPENDENCIES
