+++ src/geode/io/mesh/dxf_input.cpp
@@ -21,7 +21,7 @@
  *
  */
 
-#include <geode/io/mesh/private/dxf_input.h>
+#include <geode/io/mesh/internal/dxf_input.h>
 
 #include <geode/geometry/nn_search.h>
 #include <geode/geometry/point.h>
@@ -29,19 +29,18 @@
 #include <geode/mesh/builder/polygonal_surface_builder.h>
 #include <geode/mesh/core/polygonal_surface.h>
 
-#include <geode/io/mesh/private/assimp_input.h>
+#include <geode/io/mesh/internal/assimp_input.h>
 
 namespace geode
 {
-    namespace detail
+    namespace internal
     {
         std::unique_ptr< PolygonalSurface3D > DXFInput::read(
             const MeshImpl& /*unused*/ )
         {
-            geode::detail::AssimpMeshInput< geode::PolygonalSurface3D > reader{
-                filename()
-            };
+            geode::internal::AssimpMeshInput< geode::PolygonalSurface3D >
+                reader{ filename() };
             return reader.read_file();
         }
-    } // namespace detail
+    } // namespace internal
 } // namespace geode
