+++ src/geode/io/mesh/smesh_curve_input.cpp
@@ -21,7 +21,7 @@
  *
  */
 
-#include <geode/io/mesh/private/smesh_curve_input.h>
+#include <geode/io/mesh/internal/smesh_curve_input.h>
 
 #include <fstream>
 
@@ -34,21 +34,21 @@
 #include <geode/mesh/builder/edged_curve_builder.h>
 #include <geode/mesh/core/edged_curve.h>
 
-#include <geode/io/mesh/private/smesh_input.h>
+#include <geode/io/mesh/internal/smesh_input.h>
 
 namespace
 {
     class SMESHCurveInputImpl
-        : public geode::detail::SMESHInputImpl< geode::EdgedCurve3D,
+        : public geode::internal::SMESHInputImpl< geode::EdgedCurve3D,
               geode::EdgedCurveBuilder3D,
               2 >
     {
     public:
         SMESHCurveInputImpl(
             std::string_view filename, geode::EdgedCurve3D& curve )
-            : geode::detail::SMESHInputImpl< geode::EdgedCurve3D,
-                geode::EdgedCurveBuilder3D,
-                2 >( filename, curve )
+            : geode::internal::SMESHInputImpl< geode::EdgedCurve3D,
+                  geode::EdgedCurveBuilder3D,
+                  2 >( filename, curve )
         {
         }
 
@@ -63,7 +63,7 @@ namespace
 
 namespace geode
 {
-    namespace detail
+    namespace internal
     {
         std::unique_ptr< EdgedCurve3D > SMESHCurveInput::read(
             const MeshImpl& impl )
@@ -73,5 +73,5 @@ namespace geode
             reader.read_file();
             return curve;
         }
-    } // namespace detail
+    } // namespace internal
 } // namespace geode
