+++ src/geode/io/mesh/triangle_output.cpp
@@ -21,7 +21,7 @@
  *
  */
 
-#include <geode/io/mesh/private/triangle_output.h>
+#include <geode/io/mesh/internal/triangle_output.h>
 
 #include <fstream>
 #include <string>
@@ -94,7 +94,7 @@ namespace
 
 namespace geode
 {
-    namespace detail
+    namespace internal
     {
         std::vector< std::string > TriangleOutput::write(
             const TriangulatedSurface2D& surface ) const
@@ -110,5 +110,5 @@ namespace geode
             return { std::move( node_file ), std::move( ele_file ),
                 std::move( neigh_file ) };
         }
-    } // namespace detail
+    } // namespace internal
 } // namespace geode
