+++ src/geode/io/model/CMakeLists.txt
@@ -35,9 +35,9 @@ add_geode_library(
         "detail/vtm_section_output.h"
         "detail/vtm_brep_output.h"
     PRIVATE_HEADERS
-        "private/msh_input.h"
-        "private/msh_output.h"
-        "private/svg_input.h"
+        "internal/msh_input.h"
+        "internal/msh_output.h"
+        "internal/svg_input.h"
     PRIVATE_DEPENDENCIES
         Async++
         OpenGeode::basic
